Moved initialization code for enemy and bullet into separate files enemy.c and bullet.c respectively. Added some comments to brief on function use. Added new utility.c program for useful misc functions such as generating random positions on the screen. Modified collider_aabb.h to use float variables for width and height instead of integer variables.

+++ CProcessing/collider_aabb.h
@@ -1,5 +1,5 @@
 #pragma once
 struct Collider_AABB {
-	int width;
-	int height;
+	float width;
+	float height;
 };
